% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03clusters.R
\docType{class}
\name{clustersClass-class}
\alias{clustersClass-class}
\alias{clustersClass}
\title{Aggregate anchorages into clusters}
\description{
Aggregate anchorages into clusters
}
\section{Fields}{

\describe{
\item{\code{anchorages}}{ANY Object of class poins2Matrix}

\item{\code{deltaR}}{numeric Number of cells that must be traversed to cover 4 km in latitude}

\item{\code{nn}}{numeric Number of valid anchorages}

\item{\code{groupID}}{Matrix Sparse matrix indicating the cluster ID for each anchorage}

\item{\code{rIdx}}{numeric Row indices (in groupID) of valid anchorages}

\item{\code{cIdx}}{numeric Column indices (in groupID) of valid anchorages}

\item{\code{nClust}}{numeric Number of clusters found}

\item{\code{earthGeo}}{ANY Object of class earthGeoClass}

\item{\code{verbose}}{Output progress to console}
}}


