% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01earthGeo.R
\docType{class}
\name{earthGeoClass-class}
\alias{earthGeoClass-class}
\alias{earthGeoClass}
\title{Calculate how many cells in longitude should be used, so that the search square is roughly 0.1 * 0.1 km^2}
\description{
Calculate how many cells in longitude should be used, so that the search square is roughly 0.1 * 0.1 km^2
}
\section{Fields}{

\describe{
\item{\code{earthCircumfKm}}{Circumference of the Earth, in km}

\item{\code{length1DegLatKm}}{Length of one degree of latitude, in km}

\item{\code{resKm}}{Desired spatial resolution, in km}

\item{\code{resDeg}}{Meridional spatial resolution, in degrees}

\item{\code{nlat}}{Number of latitude cells that go from 90N to 90S}

\item{\code{nlon}}{Number of longitude cells that go from 180W to 180E}

\item{\code{lat}}{Array of latitudes (in decreasing order)}

\item{\code{lon}}{Array of longitudes (in increasing order)}
}}


