% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02points2Matrix.R
\docType{class}
\name{points2MatrixClass-class}
\alias{points2MatrixClass-class}
\alias{points2MatrixClass}
\title{Convert a dataset of points into a sparse matrix, for faster downstream computations}
\description{
Convert a dataset of points into a sparse matrix, for faster downstream computations
}
\section{Fields}{

\describe{
\item{\code{id}}{character ID of point}

\item{\code{country}}{character Country of point, if available}

\item{\code{valid}}{logical Is point valid or redundant?}

\item{\code{mat}}{Matrix Sparse matrix with 100 m resolution at the equator}

\item{\code{nlat}}{Number of latitude cells that go from 90N to 90S}

\item{\code{nlon}}{Number of longitude cells that go from 180W to 180E}

\item{\code{idx}}{numeric Total number of valid IDs}

\item{\code{delta}}{numeric Meridional spatial resolution, in degrees}

\item{\code{verbose}}{Output progress to console}
}}


