% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04polygons.R
\docType{class}
\name{polygonsClass-class}
\alias{polygonsClass-class}
\alias{polygonsClass}
\title{Convert clusters of anchorages into polygons}
\description{
Convert clusters of anchorages into polygons
}
\section{Fields}{

\describe{
\item{\code{allPolygons}}{list List of all polygons}

\item{\code{earthGeo}}{ANY Object of class earthGeoClass}

\item{\code{clusters}}{ANY Object of class ClustersClass}

\item{\code{centroids}}{data.frame Longitude and latitude of polygon centroids}

\item{\code{verbose}}{Output progress to console}
}}


